+++ src/components/ListGarmentsContainer.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react';
-import {StyleSheet, Text, View} from 'react-native';
+import {StyleSheet, View} from 'react-native';
 
-import { createGarmentAPI, listGarmentsAPI } from '../scripts/garment-api-calls.js';
-import { Button, List, ListItem, Icon } from 'react-native-elements'
+import { listGarmentsAPI } from '../scripts/garment-api-calls.js';
+import { List, ListItem } from 'react-native-elements'
 
 export default class ListGarmentsContainer extends Component {
   constructor(props)  {
@@ -15,7 +15,7 @@ export default class ListGarmentsContainer extends Component {
   componentDidMount() {
     listGarmentsAPI()
       .then(result => {
-        console.log("component did mount" ,result.data.listGarments.items);
+        // console.log("ListGarmentsContainer" , result.data.listGarments.items);
         this.setState({
           garments: result.data.listGarments.items
         });
@@ -26,8 +26,6 @@ export default class ListGarmentsContainer extends Component {
   }
 
   render() {
-    console.log('render:', this.state.garments)
-
     return (
       <View>
         <List containerStyle={{marginBottom: 20}}>
